Rename to "rushlight"
