+++ README.md
@@ -1,14 +1,14 @@
-# cm-collab
+# 🕯️ Rushlight
 
-**Collaborative text editor on your own infrastructure: just Redis and a
-database.**
+_Make collaborative text editors on your own infrastructure: just Redis and a
+database._
 
 Supports multiple real-time documents, with live cursors. Based on
-[CodeMirror](https://codemirror.net/)'s operational transformation
-[extension](https://codemirror.net/examples/collab/), so all changes are
-resolved by server code. Also, it's designed to be as easy to integrate as
+[CodeMirror 6](https://codemirror.net/) and
+[operational transformation](https://codemirror.net/examples/collab/), so all
+changes are resolved by server code. It's designed to be as easy to integrate as
 possible (read: boring). The backend is stateless, and _you can bring your own
-transport_; even a single HTTP handler is enough!
+transport_; even a single HTTP handler is enough.
 
 Unlike most toy examples, this library supports persistence in any durable
 database you choose. Real-time updates are replicated by Redis, with automatic
@@ -19,9 +19,9 @@ An experiment by [Eric Zhang](https://www.ekzhang.com/), author of
 
 ## Motivation
 
-Real-time collaborative editing can be embedded in many web applications, where
-you generally already have a database. However, most libraries are unsuitable
-because they:
+Let's say you're writing a web application. You already have a database, and you
+want to add real-time collaborative editing. However, most libraries are
+unsuitable because they:
 
 - Require proprietary gadgets
 - Are not flexible enough, e.g., to customize appearance
@@ -41,20 +41,20 @@ Install the client and server packages.
 
 ```bash
 # client
-npm install cm-collab
+npm install rushlight
 
 # server
-npm install cm-collab-server
+npm install rushlight-server
 ```
 
 On the frontend, create a `CollabClient` object and attach it to your CodeMirror
 instance via extensions.
 
 ```ts
-import { CollabClient } from "cm-collab";
 import { EditorView } from "codemirror";
+import { CollabClient } from "rushlight";
 
-const collab = await CollabClient.of({
+const rushlight = await CollabClient.of({
   async connection(message) {
     // You can use any method to send messages to the server. This example
     // executes a simple POST request.
@@ -73,8 +73,8 @@ const collab = await CollabClient.of({
 const view = new EditorView({
   extensions: [
     // ...
-    collab,
-    collab.presence, // Optional, if you want to show remote cursors.
+    rushlight,
+    rushlight.presence, // Optional, if you want to show remote cursors.
   ],
   // ...
 });
@@ -87,10 +87,10 @@ string and a persistent database for document storage.
 The example below is with `express`, but you can use any framework.
 
 ```ts
-import { Checkpoint, CollabServer } from "cm-collab-server";
 import express from "express";
+import { Checkpoint, CollabServer } from "rushlight-server";
 
-const collab = await CollabServer.of({
+const rushlight = await CollabServer.of({
   redisUrl: process.env.REDIS_URL || "redis://localhost:6473",
   async loadCheckpoint(id: string): Promise<Checkpoint> {
     // ... Load the document from your database.
@@ -101,14 +101,14 @@ const collab = await CollabServer.of({
   },
 });
 
-collab.compactionTask(); // Run this asynchronously.
+rushlight.compactionTask(); // Run this in the background.
 
 const app = express();
 
 app.post("/doc/:id", express.json(), async (req, res) => {
   const id = req.params.id;
   try {
-    res.json(await collab.handle(id, req.body));
+    res.json(await rushlight.handle(id, req.body));
   } catch (e: any) {
     console.log("Failed to handle user message:", e.toString());
     res.status(400).send(e.toString());
@@ -159,3 +159,17 @@ npm start -w=app
 ```
 
 Listens on port 6471 by default, or the `PORT` environment variable if set.
+
+## Why the name?
+
+It comes from this quote, and the fact that rushlights are a type of makeshift
+candle; you make do with what you have.
+
+> “Early Sunday morning, Natasha and I lit a candle, looked in the mirror … They
+> say you can see your future in the long row of candles, stretching back and
+> back and back, into the depths of the mirror.”
+>
+> ”I see nothing but the candle in the mirror. No visions of the future. So lost
+> and alone.”
+>
+> ―Dave Malloy, _Natasha, Pierre & The Great Comet of 1812_
