+++ app/package.json
@@ -22,8 +22,8 @@
     "@codemirror/search": "^6.5.0",
     "@codemirror/state": "^6.2.1",
     "@codemirror/view": "^6.13.2",
-    "cm-collab": "*",
-    "cm-collab-server": "*",
+    "rushlight": "*",
+    "rushlight-server": "*",
     "express": "^4.18.2",
     "nanoid": "^4.0.2",
     "postgres": "^3.3.5",
